devop: move company layout under core
