+++ src/core/viewDefault/components/companyLayout/CompanyCommunity.vue
